BACKEND/requirements.txt
fastapi==0.73.0
motor==2.5.1
pydantic[email]~=1.9.0
python-decouple==3.6
uvicorn==0.17.4
python-dotenv~=0.19.2
pymongo~=3.12.3
requests~=2.27.1
python-dateutil~=2.8.2